upgraded local redis in docker to 7.2 which matches RedisLabs version.
added singleton to load env variables from file based on NODE_ENV value with override.
this allows loading prod values easily when running locally.
loading env values now in prismaClient & redisClient which will override values when it finds the file.
adding logs to when it is able to get a lock and when not