added default text for intro.
uses variables in intro. it assumes that contact has profile, company etc.

on intro create page looking up contact profile etc. and passing to form.

made prospect search be inner join as we assume that a prospect has profile, experience and company profile.

prospect serach now ensures that linkedInUrl is present on profile.

added index on linkedInUrl for personProfile as we are checking for it.