built credits system.
added service to add credits to user. after adding credits it takes intros of the user which are in pending credits state and moves them to approve state while subtracting credits for that. the number of intros taken is equal to credits+1.

updated approve intro action. it first updats the contact message if there.
it then moves the intro to PC if requester has no credits or approved if requester had money.
it then gives 1 credit to facilitator and takes 1 credit from requester.

refactored out create intro button. thoughts were to disable it to prevent fraud, but not doing that for now.

when intro is created we send an email saying it is in pending approval

added component to show credits badge.

added logic documentation in canStateChange.

added job which sends emails of all approved intros and then marks them as sent.

added email template which is sent when an intro is pending approval.

added service which make intro approved.
added service which makes intro pending credit.

showing credits in navigation bar next to user avatar.

added alert for negative balance.

home page also has credits for debugging.

prospects table is now async component which gets the user from session. it uses the user to show negative balance alert. it also sends the user down to createIntroButton in case it wants to disalbe itself. but for now there is no disabling.

added service to remove credits.

added integer credits with default of 0.

added service which takes all intros in approved state and calls sendEmail service while passing the intro so it can be marked as sent.

added cron task to send emails at 11am