showing dialog to confirm that introduction has been created.
no longer redirecting after intro is created in an action. it is expected for the client side to redirect if there is no error.

in intro create form now storing when it was submitted. then using that along with no error state to show intro confirmation dialog. submitting form with client side handler so that we can set submittedAt first. it loads the intro creation confirmation dialog.

updated shadncui dialog to have option to hide close button.

 created intro confirmation dialog which uses router to go back when use clicks continue. it is controlled component with no close.

 showing new first tab on intros page with intros which are pending approval

 in prospects page building intros pending users approval and giving it to the third tab.