added concept of super users.
have constant of super users.
they see super page in navigation.
on super page all users are listed with form to sign in as them.

added action to sign with credentials which only works if logged in user is a super user.
it passes user id which needs to be impersonated.

added crednetials as a provider which just logs you in. this it to only be called by super users.