fixing filters form auto submit without running many times and save filters not updating form.
enabled hook to auto submit the form and replace the url when any of the items in state change.

made saved filters list links be actual anchor tags over link. this way they refresh the page. this causes the fitlers form to take values from the url.
when i use a link, it was a shallow refresh, so the component wasn't rebuilt, which means that i have to sync url to state and it either caused loops or timing issues as the replace is async.

made MyDropDown which is built on top of react-select use value instead of default value making them controlled components.

added some test components to see how replace, effect, state & url syncing plays out.

removed key from filtersForm. this was causing it to be rebuilt whenever searchString changed and remain the same if it didnt change. this meant when we add something to it, url changed and thus it changed and we lost cursor.

also made options limit to 50 from 5 for job titles.