updated profile page to have a profileImageForm
build a profileImageForm which shows the profile image and has form to update it.

added profileImageName field to store image uuid name on the user model

added action to take file, upload it to s3 with uuid and update user with that uuid for profileImageName

added helper to get s3 url, this gives url with cloudfront and does not hash the key.

added migration for profileImageName