no longer prompting for consent with metaKey aka command is pressed.
google sign in action is looking for metaKey in formData and will remove prompt property when it finds it.

sign in form no longer calls action directly. it instead has an on click handler which prevents submision, then builds formData and then calls action itself with metaKey set based on the event object.