prospects can be marked as wanted.
updated template which asks contact permission. it now has a pre-built blurb.

after fetching prospects, we also fetch if any of them are wanted by the user.

in prospects table showing a form so that user can mark if they want the prospect or not.

added wantedContact table with user now having wantsToMeet and contact having wantedBy relations.

added action to create or delete wantedContact relation.

added want to meet and want to not meet form.