checking if email is valid before sending it.
added column emailCheckPassed with default true value to contacts.
now only fetching records which have true for this column.

added service which tells if contact's email is valid. if invalid it says so. if valid, then it checks and updates db incase it is invalid.

when calling processUserForAutoProspecting job, now sending 3 attempts, so in case it throws an exception it can be retried.

processUserForAutoProspecting job now calls to check if prospects email is good, if not it throws error so it can be retried.

wrote verifyEmail.ts service which can call zerobounce api.