added competition table with  llm service to figure out if 2 users are competitive or not.
updated fetcher services to not fetch from facilitators who are competing with each other.
added service to get competitors of a user where the user is the initiator and the competitor is the receiver.
calling setup competitors on a user when a user is manually linked to a profile.
when a user is onboarded we wait up to 5 mins for apollo enrich to finish and then at end setup the competitors of that user.
added a page to show competitors of a user.
added competion table.
added service to setup competitors on all users.
added service to save competitors of a user.
added link to show competitors.