updated auto prospecting to happen for ecah group.
process user for auto prospecting, fetches groups of that user and then for each group it runs code to find best intro and make intro. in the end it returns array of intros made.

when fetching contact from email, also taking in group to limit contacts to users who are member of that group.

updated fetch contacts from icp to take group and pass that along when fetching contacts.

updated fetch wanted contacts to also take group and return only is user is member of that group.

fetch best contacts takes in group and passes to finders.