group now have manage so members can be approved.
when group is created, the creator is added to it as approved.
added action, to delete a group. the user has to be the creator to be able to delete it.
added dialog to show delete group form.
added form to delete group.
updated groups list page to show manage, approved status badge.
added approved to membership with default false.
in initial setup script adding all existing users to platform group as approved.
in action to setup membership, if user is leaving then their membership is deleted.
added action so membership can be updated
added form to updateMembership