signup page adds you to platform group by default or the group you desired by query string.
sign in with google action accepts group name from front end and adds it to the redirect url as its query string.

sign in page takes groupName from the url query string. it displays that group to the user informing where they are headed.

the home page takes in group name from the url query string. if it has a group it setups membership to it if one is not already present.

added service to setup intial membership. it takes a group name and connects user to it if it is not already connected.

sign in with google form takes group name from url or defaults to platform group. it sends that to the sign in action.