storing contacts in pinecone.
using new namesapce contacts
also storing metadata userId with id being actual contact id.
when llm description is added on a person profile, we will get all associated contacts and save them to pinecone.
added service to find contacts on user, this limits the results to contacts of that user.